+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 Change Log
 ==========
 
+Version 1.4.2 *(2020-04-15)*
+----------------------------
+* Fix: Perform click method was not triggering the spinner.
+* Fix: Set ime options to never show keyboard.
+
 Version 1.4.1 *(2019-12-27)*
 ----------------------------
 * Fix: Issue where toString was not invoked for null items.